one line to give the program's name and a brief description Copyright 2015 yourname
Licensed under the "THE BEER-WARE LICENSE" (Revision 42):
yourname wrote this file. As long as you retain this notice you
can do whatever you want with this stuff. If we meet some day, and you think
this stuff is worth it, you can buy me a beer or coffee in return